More controller methods

+++ QuantumSlots/src/Service/RewardService.java
@@ -3,7 +3,7 @@ package Service;
 import Model.Potentials;
 
 public class RewardService {
-    public int getReward(Potentials potential, float selector1, float selector2, float time, int bet){
+    public int getReward(Potentials potential, float selector1, float selector2, double time, int bet){
         switch (potential){
             case InfiniteSquareWell:
                 return -1;
